java compiler: infer map generics

+++ tests/machine/x/java/var_assignment.java
@@ -2,7 +2,7 @@ import java.util.*;
 public class Main {
 	static int x = 1;
 	public static void main(String[] args) {
-	x = 2;
+	x = (int)(2);
 	System.out.println(x);
 	}
 }
